IsPrimeX
IsPrimeX method should check whether the value sent to the method is prime or not. If the value is prime, then it will check whether the digit sum of value is prime or not. It will repeat again until the last value has a 1 digit. If the value will not be Prime value in any step, then related method should return false.